Type d

+++ moveInformation/src/main/java/com/netflix/dto/ProductionDTO/ProductionRegisterDTO.java
@@ -2,5 +2,5 @@ package com.netflix.dto.ProductionDTO;
 
 public record ProductionRegisterDTO (
         long allWorkers,
-        long allHectares
+        double allHectares
 ) {}
